refactor(ai): estrai il prompt Patient Insight
